pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.10.6",
    "httpx>=0.27.2",
    "matplotlib>=3.9.2",
    "mplfinance>=0.12.10b0",
    "openai>=1.50.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic>=2.9.2",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "questionary>=2.0.1",
    "rich>=13.8.1",
    "yfinance>=0.2.43",
]

[tool.pyright]
venv = "."
venvPath = ".venv"
